created unified API for building a ThreeScene, independent of R3F and Three.js
