+++ packages/core/src/Profiles/Scene/GLTFJson.ts
